requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0  # For Python 3.8 compatibility
flask>=2.0.0  # For REST API
flask-cors>=3.0.0  # For CORS support
websockets>=10.0  # For WebSocket server

# Optional dependencies
gradio>=3.50.0  # For web interface
langgraph>=0.0.15  # For LangGraph agent
pydantic>=2.0.0  # For data validation

# For development
pytest>=7.0.0  # For testing
black>=23.0.0  # For code formatting

nbformat>=5.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
matplotlib>=3.0.0
networkx>=3.0.0
numpy>=1.24.0
pyquaternion>=0.9.9
reachy2-sdk>=1.0.0 